+++ src/test/converter2/issues/gh2401.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Comment for module
+ * @module foo:bar
+ */
+declare module "foo:bar" {
+    export function doSomething(a: any): void;
+}